# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: M.P. Warehousing & Logistics Corporation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change user password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/approvals:
    get:
      operationId: listApprovals
      tags: [approvals]
      summary: List all approvals (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of approvals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillApproval"
  /v1/approvals/approve/{billId}:
    post:
      operationId: approveBill
      tags: [approvals]
      summary: Approve a bill (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: billId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveBillRequest"
      responses:
        "200":
          description: Bill approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillDetail"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/approvals/reject/{billId}:
    post:
      operationId: rejectBill
      tags: [approvals]
      summary: Reject a bill (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: billId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectRequest"
      responses:
        "200":
          description: Bill rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillDetail"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/approvals/versions/{versionId}/approve:
    post:
      operationId: approveVersion
      tags: [approvals]
      summary: Approve a bill version/edit request (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: versionId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Version approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillVersion"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/approvals/versions/{versionId}/reject:
    post:
      operationId: rejectVersion
      tags: [approvals]
      summary: Reject a bill version/edit request (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: versionId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectRequest"
      responses:
        "200":
          description: Version rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillVersion"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/approvals/pending-versions:
    get:
      operationId: listPendingVersions
      tags: [approvals]
      summary: List all pending version/edit requests (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of pending versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillVersion"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    ChangePasswordRequest:
      type: object
      properties:
        current_password:
          type: string
        new_password:
          type: string
      required:
        - current_password
        - new_password
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, operator]
        branch_name:
          type: ["string", "null"]
        district_name:
          type: ["string", "null"]
        mobile_number:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - created_at
...
    Commodity:
      type: object
      properties:
        id:
          type: integer
        crop_name:
          type: string
        crop_year:
          type: string
        per_bag_per_month:
          type: number
      required:
        - id
        - crop_name
        - crop_year
        - per_bag_per_month
...
    Depositor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        gst_no:
          type: ["string", "null"]
        total_gst:
          type: ["number", "null"]
      required:
        - id
        - name
...
    Bill:
      type: object
      properties:
        id:
          type: integer
        serial_no:
          type: integer
        created_by:
          type: integer
        district:
          type: ["string", "null"]
        branch_name:
          type: ["string", "null"]
        godown_name:
          type: ["string", "null"]
        bill_no:
          type: ["string", "null"]
        commodity_id:
          type: integer
        crop_year:
          type: ["string", "null"]
        financial_year:
          type: ["string", "null"]
        month_year:
          type: ["string", "null"]
        rate_per_bag:
          type: ["number", "null"]
        opening_balance:
          type: ["integer", "null"]
        closing_balance:
          type: ["integer", "null"]
        received_bags:
          type: ["integer", "null"]
        issue_bags:
          type: ["integer", "null"]
        reserve_bags:
          type: ["integer", "null"]
        chargeable_bags:
          type: ["integer", "null"]
        total_charge:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        is_locked:
          type: boolean
        created_at:
          type: string
          format: date-time
      required:
        - id
        - serial_no
        - created_by
        - commodity_id
        - status
        - is_locked
        - created_at
    BillDetail:
      type: object
      properties:
        id:
          type: integer
        serial_no:
          type: integer
        created_by:
          type: integer
        district:
          type: ["string", "null"]
        branch_name:
          type: ["string", "null"]
        godown_name:
          type: ["string", "null"]
        bill_no:
          type: ["string", "null"]
        commodity_id:
          type: integer
        crop_year:
          type: ["string", "null"]
        financial_year:
          type: ["string", "null"]
        month_year:
          type: ["string", "null"]
        rate_per_bag:
          type: ["number", "null"]
        opening_balance:
          type: ["integer", "null"]
        closing_balance:
          type: ["integer", "null"]
        received_bags:
          type: ["integer", "null"]
        issue_bags:
          type: ["integer", "null"]
        reserve_bags:
          type: ["integer", "null"]
        chargeable_bags:
          type: ["integer", "null"]
        total_charge:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        is_locked:
          type: boolean
        created_at:
          type: string
          format: date-time
        creator:
          $ref: "#/components/schemas/User"
        commodity:
          $ref: "#/components/schemas/Commodity"
        approval:
          oneOf:
            - $ref: "#/components/schemas/BillApproval"
            - type: "null"
        versions:
          type: array
          items:
            $ref: "#/components/schemas/BillVersion"
      required:
        - id
        - serial_no
        - created_by
        - commodity_id
        - status
        - is_locked
        - created_at
        - creator
        - commodity
        - versions
...
    BillVersion:
      type: object
      properties:
        id:
          type: integer
        bill_id:
          type: integer
        data_json:
          type: object
          additionalProperties: true
        version_type:
          type: string
          enum: [edit, delete]
        status:
          type: string
          enum: [pending, approved, rejected]
        created_by:
          type: integer
        created_at:
          type: string
          format: date-time
        creator:
          $ref: "#/components/schemas/User"
        bill:
          oneOf:
            - $ref: "#/components/schemas/Bill"
            - type: "null"
      required:
        - id
        - bill_id
        - data_json
        - version_type
        - status
        - created_by
        - created_at
        - creator
...
    BillApproval:
      type: object
      properties:
        id:
          type: integer
        bill_id:
          type: integer
        depositor_id:
          type: ["integer", "null"]
        pass_amount:
          type: ["number", "null"]
        payment_method:
          type: ["string", "null"]
        neft_no:
          type: ["string", "null"]
        remark:
          type: ["string", "null"]
        remark_image_url:
          type: ["string", "null"]
        approved_by:
          type: ["integer", "null"]
        approved_at:
          type: ["string", "null"]
          format: date-time
        depositor:
          oneOf:
            - $ref: "#/components/schemas/Depositor"
            - type: "null"
        approver:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
      required:
        - id
        - bill_id
    ApproveBillRequest:
      type: object
      properties:
        depositor_id:
          type: integer
        pass_amount:
          type: number
        payment_method:
          type: string
        neft_no:
          type: string
        remark:
          type: string
        remark_image_url:
          type: string
      required:
        - depositor_id
        - pass_amount
        - payment_method
    RejectRequest:
      type: object
      properties:
        reason:
          type: string
      required:
        - reason
...