# lib/api-zod/src/generated/types/billVersion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * M.P. Warehousing & Logistics Corporation API
 * OpenAPI spec version: 0.1.0
 */
import type { Bill } from "./bill";
import type { BillVersionDataJson } from "./billVersionDataJson";
import type { BillVersionStatus } from "./billVersionStatus";
import type { BillVersionVersionType } from "./billVersionVersionType";
import type { User } from "./user";

export interface BillVersion {
  id: number;
  bill_id: number;
  data_json: BillVersionDataJson;
  version_type: BillVersionVersionType;
  status: BillVersionStatus;
  created_by: number;
  created_at: Date;
  creator: User;
  bill?: Bill | null;
}