# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * M.P. Warehousing & Logistics Corporation API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  branch_name?: string | null;
  /** @nullable */
  district_name?: string | null;
  /** @nullable */
  mobile_number?: string | null;
  created_at: Date;
}